csvの読み込みが間違っていたため変更

+++ FantaRhythm_v2/Character.cpp
@@ -17,8 +17,8 @@ Character::Character(CharacterSubject* csubject, const FilePath& jobname,const C
 	name = csv.get<String>(row, 2);
 	hp = csv.get<int>(row, 3);
 	power = csv.get<int>(row,4);
-	args1 = csv.get<int>(row, 5);
-	args2 = csv.get<int>(row, 6);
+	args1 = csv.get<double>(row, 5);
+	args2 = csv.get<double>(row, 6);
 	//ƒLƒƒƒ‰‰æ‘œ‚Ì“Ç‚İ‚İ
 	TextureAsset::Register(name,U"resources/images/character/"+name+U".png");
 	TextureAsset::Preload(name);
