+++ FantaRhythm_v2/Character.h
@@ -46,7 +46,7 @@ private:
 	String name;		//キャラの名前
 	int hp;				//ヒットポイント
 	int power;			//攻撃力ベース
-	int args1, args2;	//汎用値
+	double args1, args2;	//汎用値
 	double initx, inity;//基本位置
 	double x, y;		//現在位置
 	int framecnt;		//上下移動に使うフレームカウント
